backport of commit 356ea87e00fdf310ad5d107faadd58dad5f2b079
