+++ .changelog/23432.txt
@@ -1,0 +1,3 @@
+```release-note:bug
+template: Fix template rendering on Windows
+```
